requirements.txt
python-telegram-bot==13.7
Pillow==9.0.0
python-dotenv==0.19.0
requests==2.26.0
urllib3==1.26.7